requirements.txt
fastapi
mangum
python-dotenv
langchain
langchain-community
langchain-openai
langchain-groq
faiss-cpu
pypdf
uvicorn
pydantic








